backend/pyproject.toml
[project]
name = "clear-vision"
version = "0.1.0"
description = "Clear vision backend"
authors = [{ name = "Raphael Ramos", email = "raphael.opensource@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "torch (>=2.9.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "accelerate (>=1.11.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "dependency-injector (>=4.48.2,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
packages = [{ include = "clear_vision" }]
